+++ packages/google-vertex/src/google-vertex-config.ts
@@ -1,0 +1,8 @@
+import { FetchFunction, Resolvable } from '@ai-sdk/provider-utils';
+
+export interface GoogleVertexConfig {
+  provider: string;
+  baseURL: string;
+  headers: Resolvable<Record<string, string | undefined>>;
+  fetch?: FetchFunction;
+}