+++ packages/openai-compatible/src/openai-compatible-provider.ts
@@ -4,7 +4,11 @@ import {
   NoSuchModelError,
   ProviderV1,
 } from '@ai-sdk/provider';
-import { FetchFunction, withoutTrailingSlash } from '@ai-sdk/provider-utils';
+import {
+  FetchFunction,
+  loadApiKey,
+  withoutTrailingSlash,
+} from '@ai-sdk/provider-utils';
 import { OpenAICompatibleChatLanguageModel } from './openai-compatible-chat-language-model';
 import { OpenAICompatibleChatSettings } from './openai-compatible-chat-settings';
 import { OpenAICompatibleCompletionLanguageModel } from './openai-compatible-completion-language-model';
@@ -45,23 +49,31 @@ export interface OpenAICompatibleProvider<
 
 export interface OpenAICompatibleProviderSettings {
   /**
-   Base URL for the API calls.
+Base URL for the API calls.
    */
   baseURL?: string;
 
   /**
-   Custom headers to include in the requests.
+API key for authenticating requests. If specified, adds an `Authorization`
+header to request headers with the value `Bearer <apiKey>`. This will be added
+before any headers potentially specified in the `headers` option.
+   */
+  apiKey?: string;
+
+  /**
+Optional custom headers to include in requests. These will be added to request headers
+after any headers potentially added by use of the `apiKey` option.
    */
   headers?: Record<string, string>;
 
   /**
-   Custom fetch implementation. You can use it as a middleware to intercept requests,
-   or to provide a custom fetch implementation for e.g. testing.
+Custom fetch implementation. You can use it as a middleware to intercept requests,
+or to provide a custom fetch implementation for e.g. testing.
    */
   fetch?: FetchFunction;
 
   /**
-   Provider name.
+Provider name.
    */
   name?: string;
 }
@@ -97,10 +109,15 @@ export function createOpenAICompatible<
     fetch?: FetchFunction;
   }
 
+  const getHeaders = () => ({
+    ...(options.apiKey && { Authorization: `Bearer ${options.apiKey}` }),
+    ...options.headers,
+  });
+
   const getCommonModelConfig = (modelType: string): CommonModelConfig => ({
     provider: `${providerName}.${modelType}`,
     url: ({ path }) => `${baseURL}${path}`,
-    headers: () => options.headers ?? {},
+    headers: getHeaders,
     fetch: options.fetch,
   });