+++ packages/provider/src/image-model/v1/image-model-v1.ts
@@ -1,4 +1,5 @@
-import { JSONValue } from '../../json-value/json-value';
+import { ImageModelV1CallOptions } from './image-model-v1-call-options';
+import { ImageModelV1CallWarning } from './image-model-v1-call-warning';
 
 /**
 Image generation model specification version 1.
@@ -24,54 +25,27 @@ Provider-specific model ID for logging purposes.
   readonly modelId: string;
 
   /**
+Limit of how many images can be generated in a single API call.
+If undefined, we will max generate one image per call.
+   */
+  readonly maxImagesPerCall: number | undefined;
+
+  /**
 Generates an array of images.
    */
-  doGenerate(options: {
+  doGenerate(options: ImageModelV1CallOptions): PromiseLike<{
     /**
-Prompt for the image generation.
+Generated images as base64 encoded strings or binary data.
+The images should be returned without any unnecessary conversion.
+If the API returns base64 encoded strings, the images should be returned
+as base64 encoded strings. If the API returns binary data, the images should
+be returned as binary data.
      */
-    prompt: string;
+    images: Array<string> | Array<Uint8Array>;
 
     /**
-Number of images to generate.
+Warnings for the call, e.g. unsupported settings.
      */
-    n: number;
-
-    /**
-Size of the images to generate. Must have the format `{width}x{height}`.
-     */
-    size: `${number}x${number}` | undefined;
-
-    /**
-Additional provider-specific options that are passed through to the provider
-as body parameters.
-
-The outer record is keyed by the provider name, and the inner
-record is keyed by the provider-specific metadata key.
-```ts
-{
-  "openai": {
-    "style": "vivid"
-  }
-}
-```
-     */
-    providerOptions: Record<string, Record<string, JSONValue>>;
-
-    /**
-Abort signal for cancelling the operation.
-     */
-    abortSignal?: AbortSignal;
-
-    /**
-  Additional HTTP headers to be sent with the request.
-  Only applicable for HTTP-based providers.
-     */
-    headers?: Record<string, string | undefined>;
-  }): PromiseLike<{
-    /**
-Generated images as base64 encoded strings.
-     */
-    images: Array<string>;
+    warnings: Array<ImageModelV1CallWarning>;
   }>;
 };