+++ packages/deepseek/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @ai-sdk/deepseek
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [90fb95a]
+- Updated dependencies [e6dfef4]
+- Updated dependencies [6636db6]
+  - @ai-sdk/provider-utils@2.0.7
+  - @ai-sdk/openai-compatible@0.0.15
+
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies [19a2ce7]
+- Updated dependencies [19a2ce7]
+- Updated dependencies [43b37f7]
+- Updated dependencies [6337688]
+  - @ai-sdk/provider@1.0.4
+  - @ai-sdk/provider-utils@2.0.6
+  - @ai-sdk/openai-compatible@0.0.14
+
 ## 0.0.1
 
 ### Patch Changes