+++ packages/togetherai/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @ai-sdk/togetherai
 
+## 0.0.18
+
+### Patch Changes
+
+- Updated dependencies [90fb95a]
+- Updated dependencies [e6dfef4]
+- Updated dependencies [6636db6]
+  - @ai-sdk/provider-utils@2.0.7
+  - @ai-sdk/openai-compatible@0.0.15
+
+## 0.0.17
+
+### Patch Changes
+
+- Updated dependencies [19a2ce7]
+- Updated dependencies [19a2ce7]
+- Updated dependencies [43b37f7]
+- Updated dependencies [6337688]
+  - @ai-sdk/provider@1.0.4
+  - @ai-sdk/provider-utils@2.0.6
+  - @ai-sdk/openai-compatible@0.0.14
+
+## 0.0.16
+
+### Patch Changes
+
+- 32742bb: feat (providers): Add DeepSeek-V3 model to deepinfra/togetherai.
+
 ## 0.0.15
 
 ### Patch Changes